+++ TestCases.R
@@ -52,6 +52,8 @@ test.boston.gbm <- function() {
   instance <- Boston[370,1:13]
   ciu <- ciu.new(gbm, medv~., Boston)
   p <- ciu$ggplot.col.ciu(instance); print(p)
+  p <- ciu$ggplot.col.ciu(instance, plot.mode = "overlap"); print(p)
+  p <- ciu$ggplot.col.ciu(instance, cu.colours=NULL, plot.mode = "overlap"); print(p)
   ciu$barplot.ciu(instance, sort="CI", use.influence=TRUE)
   # See how lstat,rm,crim affect output.
   oldpar <- par(no.readonly = TRUE)
@@ -76,11 +78,13 @@ test.heart.disease.rf <- function() {
   # Random Forest with caret.
   kfoldcv <- trainControl(method="cv", number=10)
   performance_metric <- "Accuracy"
-  rf.heartdisease <- train(num~., data=heart.data, method="rf", metric=performance_metric, trControl=kfoldcv,preProcess=c("center", "scale"))
+  rf.heartdisease <- caret::train(num~., data=heart.data, method="rf", metric=performance_metric, trControl=kfoldcv,preProcess=c("center", "scale"))
   n.in <- ncol(heart.data) - 1
   instance <- heart.data[32,1:n.in]
   ciu <- ciu.new(rf.heartdisease, num~., heart.data, output.names=c("No Heart Disease","Heart Disease Present"))
   p <- ciu$ggplot.col.ciu(instance, c(1:n.in)); print(p)
+  p <- ciu$ggplot.col.ciu(instance, c(1:n.in), plot.mode = "overlap"); print(p)
+  p <- ciu$ggplot.col.ciu(instance, c(1:n.in), cu.colours=NULL, plot.mode = "overlap"); print(p)
   ciu$barplot.ciu(instance, ind.output=1, sort="CI")
   ciu$barplot.ciu(instance, ind.output=2, sort="CI")
   ciu$pie.ciu(instance, ind.output=1, sort="CI")
@@ -104,9 +108,9 @@ test.cars.UCI.rf <- function() {
   n.in <- ncol(car.data) - 1
   n.out <- 1
   # Random Forest with caret
-  kfoldcv <- trainControl(method="cv", number=10)
+  kfoldcv <- caret::trainControl(method="cv", number=10)
   performance_metric <- "Accuracy"
-  rf.carsUCI <- train(result~., data=car.data, method="rf", metric=performance_metric, trControl=kfoldcv)
+  rf.carsUCI <- caret::train(result~., data=car.data, method="rf", metric=performance_metric, trControl=kfoldcv)
   inst.ind <- 1098 # A very good one"
   instance <- car.data[inst.ind,1:6]
   ciu <- ciu.new(rf.carsUCI, formula=result~., data=car.data, vocabulary=voc)
@@ -132,8 +136,8 @@ test.cars.UCI.rf <- function() {
 # Diamonds with GBM. This is a mixed discrete/continuous input case. Takes
 # a long time for GBM to train!
 test.diamonds.gbm <- function() {
-  kfoldcv <- trainControl(method="cv", number=10)
-  diamonds.gbm <- train(price~., diamonds, method="gbm", trControl=kfoldcv)
+  kfoldcv <- caret::trainControl(method="cv", number=10)
+  diamonds.gbm <- caret::train(price~., diamonds, method="gbm", trControl=kfoldcv)
   inst.ind <- 27750 # An expensive one!
   instance <- diamonds[inst.ind,-7]
   ciu <- ciu.new(diamonds.gbm, price~., diamonds)
@@ -155,8 +159,8 @@ test.titanic.rf <- function() {
   titanic_train <- na.omit(titanic_train)
 
   # Using Random Forest here. Really bad results... Takes a while also (15 seconds?)
-  kfoldcv <- trainControl(method="cv", number=10)
-  model_rf <- train(survived ~ ., titanic_train, method="rf", trControl=kfoldcv)
+  kfoldcv <- caret::trainControl(method="cv", number=10)
+  model_rf <- caret::train(survived ~ ., titanic_train, method="rf", trControl=kfoldcv)
 
   # Example instance
   new_passenger <- data.frame(
@@ -170,6 +174,7 @@ test.titanic.rf <- function() {
   )
   ciu <- ciu.new(model_rf, survived~., titanic_train)
   p <- ciu$ggplot.col.ciu(new_passenger); print(p)
+  p <- ciu$ggplot.col.ciu(new_passenger, plot.mode = "overlap"); print(p)
   for ( i in 1:ncol(new_passenger) )
     ciu$plot.ciu(new_passenger,i,1)
   for ( i in 1:ncol(new_passenger) )
@@ -256,7 +261,6 @@ test.adult.rf <- function() {
     ciu$plot.ciu(instance[,1:n.in],i,2, n.points = 200, main = "", ylab="")
   par("cex.lab"=1)
 }
-
 
 # par(mai=c(0.8,1.2,0.4,0.2)) # Good parameters for barplot so that labels fit in.
 # par(mai=c(0.8,1.2,0.4,0.2))